Split BareKineticEnergy.h into cpp.
